tools: dummy_create: V3H: Load BL2 image to SRAM by default
This moves BL2 image to SRAM by default. There seems to be
no h/w issues with SRAM on the v2.0 SoC. Besides, the latest
ATF provides a small DTB image so that U-Boot can find out
how much memory is available. The DTB is located along with
the BL2 image which is inaccessible by non-secure mode when
located in the DRAM below 0x48000000.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>